tech(api): move GET challenges/id endpoint to evaluation BC
