+++ api/src/evaluation/routes.js
@@ -3,6 +3,7 @@ import * as assessmentRoutes from './application/assessments/index.js.js';
 import * as autonomousCoursesRoutes from './application/autonomous-courses/index.js';
 import * as badgeCriteriaRoutes from './application/badge-criteria/index.js';
 import * as badgesRoutes from './application/badges/index.js';
+import * as challengesRoutes from './application/challenges/index.js';
 import * as competenceEvaluationsRoutes from './application/competence-evaluations/index.js';
 import * as courses from './application/courses/course-route.js';
 import * as feedbacksRoutes from './application/feedbacks/index.js';
@@ -17,6 +18,7 @@ const evaluationRoutes = [
   autonomousCoursesRoutes,
   badgeCriteriaRoutes,
   badgesRoutes,
+  challengesRoutes,
   competenceEvaluationsRoutes,
   courses,
   feedbacksRoutes,
