+++ api/src/shared/routes.js
@@ -1,9 +1,8 @@
 import * as assessmentsRoutes from './application/assessments/index.js';
-import * as challengesRoutes from './application/challenges/index.js';
 import * as countryRoutes from './application/country/country-route.js';
 import * as featureToggles from './application/feature-toggles/index.js';
 import * as healthcheck from './application/healthcheck/index.js';
 
-const sharedRoutes = [healthcheck, assessmentsRoutes, challengesRoutes, featureToggles, countryRoutes];
+const sharedRoutes = [healthcheck, assessmentsRoutes, featureToggles, countryRoutes];
 
 export { sharedRoutes };
